#213: Move useClickOutside to shared hooks
