+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from "./useClickOutside";
 export * from "./useSwipe";
 export * from "./useMutationQuery";
 export * from "./useGetQuery";
